# docs/swagger.yaml
basePath: /
definitions:
...
  channel.CreateChannel:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      fm_cpe_1:
        type: string
      fm_cpe_2:
        type: string
      fm_daily_avg_1:
        type: string
      fm_daily_avg_2:
        type: string
      fm_second_eval_1:
        type: string
      fm_second_eval_2:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      tv_daily_avg_views:
        type: string
      tv_univision_number:
        type: string
      type:
        type: string
      website:
        type: string
    required:
    - image
    - name
    - type
    type: object
  channel.UpdateChannel:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      fm_cpe_1:
        type: string
      fm_cpe_2:
        type: string
      fm_daily_avg_1:
        type: string
      fm_daily_avg_2:
        type: string
      fm_second_eval_1:
        type: string
      fm_second_eval_2:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      tv_daily_avg_views:
        type: string
      tv_univision_number:
        type: string
      type:
        type: string
      website:
        type: string
    required:
    - image
    - name
    - type
    type: object
...
  utils.Base64Struct:
    properties:
      base64:
        type: string
    type: object
  utils.DefaultParam:
    properties:
      key:
        type: string
      value: {}
    type: object
  utils.FilterObj:
    properties:
      field_name:
        type: string
      field_type:
        type: string
      operation:
        type: string
      value:
        type: string
      values:
        items: {}
        type: array
    type: object
  utils.Pagination:
    properties:
      current_page_no:
        type: integer
      per_page:
        type: integer
      sort:
        type: string
      total_elements:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.RequestObj:
    properties:
      default_param:
        items:
          $ref: '#/definitions/utils.DefaultParam'
        type: array
      filter:
        items:
          $ref: '#/definitions/utils.FilterObj'
        type: array
      glob_operation:
        type: string
      page_no:
        type: integer
      per_page:
        type: integer
      sort:
        type: string
    type: object
  utils.ResponseObj:
    properties:
      data: {}
      error: {}
      pagination:
        $ref: '#/definitions/utils.Pagination'
      response_code:
        type: integer
      response_msg:
        type: string
    type: object
host: http://103.168.56.249:8080
info:
  contact: {}
  description: This is a sample API with Fiber and Swagger
  title: Ramen API
  version: "1.0"
paths:
...
  /channel:
    post:
      consumes:
      - application/json
      description: Create channel
      parameters:
      - description: CreateChannel
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/channel.CreateChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create channel
      tags:
      - Channel
  /channel/{id}:
    delete:
      consumes:
      - application/json
      description: Delete channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete channel
      tags:
      - Channel
    get:
      consumes:
      - application/json
      description: Get channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get channel
      tags:
      - Channel
    put:
      consumes:
      - application/json
      description: Update channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateChannel
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/channel.UpdateChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update channel
      tags:
      - Channel
  /channel/list:
    post:
      consumes:
      - application/json
      description: Get channel list
      parameters:
      - description: RequestObj
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get channel list
      tags:
      - Channel
...
swagger: "2.0"